requirements.txt
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
google-auth==2.13.0
idna==3.4
jproperties==2.1.1
kubernetes==24.2.0
oauthlib==3.2.2
properties==0.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.2
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
urllib3==1.26.12
websocket-client==1.4.1
